+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+v26.1.68 (2026-03-18)
+-------------------------
+ * Add messages mode to mrindex command
+
+v26.1.67 (2026-03-18)
+-------------------------
+ * Reduce elastic writer batch size to 250
+ * Tweak OS message index
+
+v26.1.66 (2026-03-18)
+-------------------------
+ * Switch contact indexing from OpenSearch to Elastic v2 index
+
 v26.1.65 (2026-03-12)
 -------------------------
  * Change some contact index fields to be keyword